requirements.txt
python-telegram-bot==22.7
pypdf2==3.0.1
openai
flask
gunicorn
